requirements.txt
# Application Moto Leboncoin Analyzer
# Python 3.10+

# Framework web
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
flask==3.0.3

# Visualisation
plotly==5.24.1

# Data processing
pandas==2.2.3
numpy==2.1.3

# Scraping
firecrawl-py==1.6.4
pydantic>=2.0.0
httpx>=0.24.0
requests==2.32.3
beautifulsoup4==4.12.3

# Utilitaires
python-dotenv==1.0.1

# Optionnel pour création d'exécutable
# pyinstaller==6.11.1